+++ modules/Student_Billing/StudentFees.php
@@ -24,6 +24,19 @@ if ( ! empty( $_REQUEST['values'] )
 		{
 			$sql = "UPDATE billing_fees SET ";
 
+			$columns['FILE_ATTACHED'] = _saveFeesFile( $id );
+
+			if ( ! $columns['FILE_ATTACHED'] )
+			{
+				unset( $columns['FILE_ATTACHED'] );
+
+				if ( empty( $columns ) )
+				{
+					// No file, and FILE_ATTACHED was the only column, skip.
+					continue;
+				}
+			}
+
 			foreach ( (array) $columns as $column => $value )
 			{
 				$sql .= DBEscapeIdentifier( $column ) . "='" . $value . "',";
@@ -42,19 +55,7 @@ if ( ! empty( $_REQUEST['values'] )
 			$fields = 'STUDENT_ID,SCHOOL_ID,SYEAR,ASSIGNED_DATE,';
 			$values = "'" . UserStudentID() . "','" . UserSchool() . "','" . UserSyear() . "','" . DBDate() . "',";
 
-			if ( isset( $_FILES['FILE_ATTACHED'] ) )
-			{
-				$columns['FILE_ATTACHED'] = FileUpload(
-					'FILE_ATTACHED',
-					$FileUploadsPath . UserSyear() . '/student_' . UserStudentID() . '/',
-					FileExtensionWhiteList(),
-					0,
-					$error
-				);
-
-				// Fix SQL error when quote in uploaded file name.
-				$columns['FILE_ATTACHED'] = DBEscapeString( $columns['FILE_ATTACHED'] );
-			}
+			$columns['FILE_ATTACHED'] = _saveFeesFile( $id );
 
 			$go = 0;